requirements.txt
numpy
scipy
progressbar2
matplotlib
numba
statsmodels
nkpack
